*: create separate namespace informers if needed
Service/pod monitors and rules resources watched by the operator
can live in different namespaces than the configmaps and secrets
associated to the Prometheus custom resources.

Before this change, we had only one namespace informer scoped to the
monitor/rule namespaces. Whenever a change of configmap or secret was
detected, the operator couldn't retrieve the associated namespace
because it didn't exist in the monitor/rule namespaces and the
reconciliation wouldn't happen. The same issue applies for Thanos Ruler
and rules.

With this change, separate namespace informers are used when the
monitors/rules and prometheus namespaces aren't identical.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>